+++ src/app/translators/AccessorSupportMapper.ts
@@ -15,16 +15,20 @@ function isGetter(obj: object, prop: string): boolean {
     return Boolean(descriptor) && Boolean(descriptor.get)
 }
 
-/**
- * Checks if `obj.prop` is a setter.
- */
-// function isSetter(obj: object, prop: string): boolean {
-//     const descriptor = Object.getOwnPropertyDescriptor(
-//         Object.getPrototypeOf(obj),
-//         prop,
-//     )
-//     return Boolean(descriptor) && Boolean(descriptor.set)
-// }
+function capitalize(source: string) {
+    source = source.replace(/[_#]/g, '')
+    return `${source.charAt(0).toUpperCase()}${source.substr(1)}`
+}
+
+function setFunc(obj: object, prop: string) {
+    const setName = `set${capitalize(prop)}`
+    return (typeof obj[setName] === 'function') ? obj[setName] : null
+}
+
+function getFunc(obj: object, prop: string) {
+    const getName = `get${capitalize(prop)}`
+    return (typeof obj[getName] === 'function') ? obj[getName] : null
+}
 
 function describeAccessor(obj: object) {
     return (prop: string): AccessorDescription => {
@@ -90,12 +94,19 @@ export class AccessorSupportMapper<T> extends ModelAutoMapper<T> {
      * working well with our custom converter.
      */
     protected _forAllMembers(destObj: any, destPropName: string, srcObj: any): void {
-        if (
-            destPropName.startsWith('_') // Private field (by convention)
-            || destPropName.startsWith('#')) { // Native private field (>= Node v12)
-                return
+        const setFn = setFunc(destObj, destPropName)
+        if (!setFn &&
+            (destPropName.startsWith('_') // Private field (by convention)
+            || destPropName.startsWith('#') // Native private field (>= Node v12)
+            )
+        ) {
+            return
         }
-        destObj[destPropName] = srcObj[destPropName]
+        if (setFn) {
+            setFn.call(destObj, srcObj[destPropName], srcObj)
+        } else {
+            destObj[destPropName] = srcObj[destPropName]
+        }
     }
 
     protected _forAllAccessors(destObj: any, srcObj: any, desc: AccessorDescription): void {
@@ -105,8 +116,8 @@ export class AccessorSupportMapper<T> extends ModelAutoMapper<T> {
                 srcVal = srcObj[desc.name]
             }
             else { // If src getter isn't available, we try method `getProp()`
-                const getFn = srcObj[`get${desc.name.toUpperCase()}`] // Eg: name => getName
-                if (typeof getFn === 'function') {
+                const getFn = getFunc(srcObj, desc.name) // Eg: name => getName
+                if (getFn) {
                     srcVal = getFn.call(srcObj) // Equiv: srcObj.getName()
                 } else {
                     // Normal property value