+++ src/app/validators/JoiExtended.ts
@@ -12,11 +12,24 @@ const bigintRule = {
     },
 
     validate(this: joi.ExtensionBoundSchema, params: any, value: any, state: joi.State, options: joi.ValidationOptions) {
-        if (typeof value !== 'bigint') {
-            // Generate an error, state and options need to be passed
-            return this.createError('genn.bigint', { v: value }, state, options)
+        let isBigInt = true
+        try {
+            // '987654321' => valid
+            // 'ABC876' => invalid
+            // Other non-bigint non-string values are all invalid
+            isBigInt = (typeof value === 'string')
+                ? (typeof BigInt(value) === 'bigint')
+                : (typeof value === 'bigint')
         }
-        return value // Everything is OK
+        catch {
+            isBigInt = false
+        }
+
+        return isBigInt
+            // Everything is OK
+            ? value
+            // Generate an error, state and options need to be passed
+            : this.createError('genn.bigint', { v: value }, state, options)
     },
 }
 
@@ -109,7 +122,12 @@ const joiExtensions: joi.Extension = {
     pre(value: any, state: joi.State, options: joi.ValidationOptions) {
         const flags: GennFlag = this['_flags']
         if (flags.bigint === true) {
-            return (options.convert ? BigInt(value) : value)
+            try {
+                return (options.convert ? BigInt(value) : value)
+            }
+            catch {
+                return value
+            }
         }
         return value // Keep the value as it was
     },