+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `JoiModelValidator.create` now accepts single param object.
 - Implemented `AccessorSupportMapper` to fix [#2](https://github.com/gennovative/micro-fleet-common/issues/2) (ModelAutoMapper should support mapping getters and setters)
 - Added static method `ValidationError.fromJoi` to fix [#3](https://github.com/gennovative/micro-fleet-common/issues/3) (ValidationError constructor should not reference joi)
+- Removed JS native `BigInt`, and using string instead.
 
 ### 1.3.1
 - Removed script "postinstall" from `package.json`.